refactor: break all 6 circular dependencies detected by madge
Baseline: 6 cycles across backend and frontend.

### audit/port.ts ↔ audit/logger.ts
Port imported AuditEntry/AuditRecord/AuditQueryOptions from logger.ts;
logger.ts imported AuditLoggerPort from port.ts.
Fix: extracted shared types to src/audit/types.ts. Port and logger both
import from types.ts. logger.ts re-exports the types for backward compat.

### prompt/port.ts ↔ prompt/store.ts
Port imported PromptTemplate/PreviewResult from store.ts; store.ts
imported TemplateStorePort from port.ts.
Fix: extracted shared types to src/prompt/types.ts. Port and store both
import from types.ts. store.ts re-exports the types for backward compat.

### orchestrator/context.ts ↔ orchestrator/retry-coordinator.ts (×3 cycles)
context.ts imported RetryCoordinator from retry-coordinator.ts;
retry-coordinator.ts imported OutcomeContext and RetryRuntimeContext from
context.ts, and transitively pulled in completion-writeback.ts and
terminal-paths.ts (which also import from context.ts).
Fix: moved RetryCoordinator interface into context.ts (where OutcomeContext
and RetryRuntimeContext are already defined). retry-coordinator.ts now
imports RetryCoordinator from context.ts alongside the other types it
already needed, and re-exports it for existing consumers.

### frontend: settings/index.ts → unified-settings-page.ts → config-view.ts → settings/index.ts
The settings barrel re-exported from unified-settings-page.ts which
imported config-view.ts which imported getValueAtPath from the barrel.
Fix: removed the unified-settings-page.ts re-exports from the barrel
(they had no external consumers — the settings page imports directly from
unified-settings-view.ts). Barrel remains the canonical entry point for
all other settings exports.

Before: 6 circular dependencies
After:  0 circular dependencies (madge --circular confirms)
